+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## master
 * Generate ES6 class 
 * Fix links to `sham-ui` repo
+* Reimplement blocks for support update from block owner component ( block owner updated -> render block usage )
+* Update `sham-ui`
 
 ## 3.2.5
 * Update dependencies